requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
SQLAlchemy==2.0.21
PyMySQL==1.1.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
bcrypt==4.0.1